Move fault insert logic to datasetDao
